// src/types/api/service/models/JournalistSrchDto.ts
export interface JournalistSrchDto {
    /**
     * 
     * @type {number}
     * @memberof JournalistSrchDto
     */
    jrnlstSrchId?: number;
    /**
     * 
     * @type {string}
     * @memberof JournalistSrchDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof JournalistSrchDto
     */
    shareCode?: string;
    /**
     * 
     * @type {string}
     * @memberof JournalistSrchDto
     */
    shareTargetCode?: string;
    /**
     * 
     * @type {number}
     * @memberof JournalistSrchDto
     */
    companyId?: number;
    /**
     * 
     * @type {number}
     * @memberof JournalistSrchDto
     */
    groupId?: number;
    /**
     * 
     * @type {string}
     * @memberof JournalistSrchDto
     */
    conditions?: string;
    /**
     * 
     * @type {UserDtoForGroup}
     * @memberof JournalistSrchDto
     */
    owner?: UserDtoForGroup;
    /**
     * 
     * @type {UserDtoForGroup}
     * @memberof JournalistSrchDto
     */
    register?: UserDtoForGroup;
    /**
     * 
     * @type {UserDtoForGroup}
     * @memberof JournalistSrchDto
     */
    updater?: UserDtoForGroup;
    /**
     * 
     * @type {Date}
     * @memberof JournalistSrchDto
     */
    regisAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof JournalistSrchDto
     */
    updateAt?: Date;
}